Add in blueprint for the AI car
Created a new blueprint for an AI car with smoother and more realistic movement than previous enemy (except currently the animation for the wheels is missing). The AI car detects the nearest spline point ahead of it and routes to that given point. For now, I added it to the FirstMap to test it out but the leaderboard and timing logic for it is not implemented yet. Also, the AI car requires a specific road parameter so we might need to duplicate one for each map to match the blueprint being used for road.